DiscoveryAgent/pyproject.toml
[project]
name = "adk_root_agent"
version = "0.1.0"
description = "ADK version of Bootstrap_Agent Assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3~=1.40.35",
    "cloudpickle==3.1.1",
    "google-adk[eval]~=1.20.0",
    "google-adk~=1.20.0",
    "google-cloud-aiplatform>=1.125.0",
    "google-cloud-dialogflow-cx~=1.42.0",
    "google-cloud-logging~=3.12.1",
    "jsonpath-ng>=1.7.0",
    "model2vec==0.6.0",
    "openai~=1.107.1",
    "openpyxl==3.1.5",
    "opentelemetry-api==1.37.0",
    "opentelemetry-sdk==1.37.0",
    "opentelemetry-exporter-gcp-logging~=1.9.0a0",
    "opentelemetry-exporter-gcp-monitoring~=1.9.0a0",
    "opentelemetry-exporter-gcp-trace~=1.9.0",
    "opentelemetry-instrumentation-google-genai~=0.3b0",
    "opentelemetry-instrumentation-vertexai>=2.0b0",
    "opentelemetry-instrumentation~=0.56b0",
    "pydantic==2.11.7",
    "pytest-asyncio>=1.1.0",
    "pytest-timeout~=2.4.0",
    "pytest-xdist~=3.3.1",
    "pytest~=8.4.1",
    "python-dotenv~=1.1.1",
    "requests~=2.32.5",
    "rouge-score~=0.1.2",
    "ruamel-yaml>=0.18.15",
    "scipy~=1.16.2",
    "tenacity~=9.1.2",
    "tqdm~=4.67.1",
    "xlsxwriter==3.2.9",
]

[dependency-groups]
dev = [
    # these three are needed for docker server for ai4cx testing envs
    "fastapi-sso==0.18.0",
    "psycopg2-binary==2.9.10",
    "python-jose[cryptography]==3.5.0",
]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]
exclude = [".venv"]

[tool.setuptools.packages.find]
where = ["."]  # list of folders that contain the packages (["."] by default)
include = ["bootstrap_agent", "bootstrap_agent.*"]  # package names should match these glob patterns (["*"] by default)
exclude = []  # exclude packages matching these glob patterns (empty by default)
namespaces = true  # to disable scanning PEP 420 namespaces (true by default)
